+++ CHANGELOG.md
@@ -7,6 +7,9 @@
       around distribution lifecycle
     - Added new `activeDistributions()` method to `DistributionLifecycleStateStore` interface with a default
       implementation
+- Event Source Improvements:
+    - `KafkaConfiguration` now auto-configures DLQ topic when relevant configuration methods are used
+    - Fixed a bug in `KafkaConfiguration` that could occur if a configuration is built with `null` client properties
 - Documentation Improvements:
     - Document new `DistributionLifecycleConfiguration` helper
     - Document previously undocumented `KafkaConfiguration` helper